με αφορμή την κοπή της πίττας, να ευχηθούμε υγεία στην γνήσια ελληνική μασωνική οικογένεια. επειδή για μικροανταλλάγματα πηδάς τη σύζυγο, πηδάς την κορούλα, όμως πηδάς και τον γιό. Αλλά πάντα την καλύτερη κωλοτρυπίδα θα την έχει ο μπαμπάς Σαχαρίδης!

+++ src/defs.h
@@ -139,7 +139,7 @@ typedef struct {
 /***** N-TREE DEFINITIONS BEGIN *****/
 
 typedef float arc_weight_t;
-typedef uint16_t arc_pointer_t;
+typedef uint32_t arc_pointer_t;
 
 /**
  * only arc sources information is kept in internal nodes,
